adds basic component folders and builds profile/header
