+++ client/src/components/Header/index.js
@@ -1,46 +1,26 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React, { useState } from 'react';
+import ProfileIcon from './ProfileIcon';
 
-import Auth from '../../utils/auth';
-
-const Header = () => {
-  const logout = (event) => {
-    event.preventDefault();
-    Auth.logout();
-  };
+const Header = ({iconUrl, onIconClick, title}) => {
+  const [searchValue, setSearchValue] = useState('');
+  const handleSearch = (e) => {
+    setSearchValue(e.target.value);
+  }
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    console.log(`Searching for: ${searchValue}`);
+    // MAKE QUERY HERE
+  }
   return (
-    <header className="bg-primary text-light mb-4 py-3 flex-row align-center">
-      <div className="container flex-row justify-space-between-lg justify-center align-center">
-        <div>
-          <Link className="text-light" to="/">
-            <h1 className="m-0">Tech Thoughts</h1>
-          </Link>
-          <p className="m-0">Get into the mind of a programmer.</p>
-        </div>
-        <div>
-          {Auth.loggedIn() ? (
-            <>
-              <Link className="btn btn-lg btn-info m-2" to="/me">
-                {Auth.getProfile().data.username}'s profile
-              </Link>
-              <button className="btn btn-lg btn-light m-2" onClick={logout}>
-                Logout
-              </button>
-            </>
-          ) : (
-            <>
-              <Link className="btn btn-lg btn-info m-2" to="/login">
-                Login
-              </Link>
-              <Link className="btn btn-lg btn-light m-2" to="/signup">
-                Signup
-              </Link>
-            </>
-          )}
-        </div>
-      </div>
+    <header>
+      <ProfileIcon iconUrl={iconUrl} onClick={onIconClick} />
+      <h1>{title}</h1>
+      <form onSubmit={handleSubmit}>
+        <input type="text" value={searchValue} onChange={handleSearch} placeholder="Search..."/>
+        <button type="submit">Search</button>
+      </form>
     </header>
   );
-};
+}
 
 export default Header;
