+++ client/src/components/LeftNav/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ProfileIcon = ({iconUrl, onClick}) => {
+  return (
+    <div onClick={onClick}>
+      <img src={iconUrl} alt="Profile Icon" />
+    </div>
+  );
+}
+
+export default ProfileIcon;
