+++ client/src/components/Template/ThoughtForm/index.js
