+++ client/src/components/Template/ThoughtList/index.js
